03-upgrading/azure-ai-agents/requirements.txt
python-dotenv==1.2.1
azure-identity==1.26.0b1
azure-ai-projects==2.0.0b2
azure-ai-agents==1.2.0b6
requests==2.32.5
